Add binary add for Int32 inputs
* Add binary add for Int32 inputs

Issue #2286

Add support for binary add for Int32 inputs.
This change only adds support for standalone version of the op,
fused op support is not implemented yet.
Sub and mul are not supported.

Int8 and Int32 can be mixed on inputs.

Op is implemented as binary sfpu op, and probably has signifiicantly
different perf profile to other binary add versions.
Perf estimates are not available yet.